Migrate Attachment tests to rspec
